Backend/requirements.txt
flask==2.0.3
werkzeug==2.0.3
flask-cors==3.0.10
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
transformers==4.41.2
google-generativeai 